Types now extracted out:
