+++ contracts/service-provider-directory/src/state/config.rs
@@ -1,14 +1,14 @@
 use cosmwasm_std::{Addr, Coin, Storage};
-use cw_storage_plus::{Index, IndexList, IndexedMap, Item, MultiIndex};
+use cw_storage_plus::Item;
 use serde::{Deserialize, Serialize};
 
 use crate::error::Result;
 
-pub const CONFIG_KEY: &str = "config";
-pub const CONFIG: Item<Config> = Item::new(CONFIG_KEY);
+const CONFIG_KEY: &str = "config";
+const CONFIG: Item<Config> = Item::new(CONFIG_KEY);
 
 #[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
-pub struct Config {
+pub(crate) struct Config {
     pub admin: Addr,
     pub deposit_required: Coin,
 }
@@ -27,7 +27,6 @@ pub(crate) fn deposit_required(store: &dyn Storage) -> Result<Coin> {
 }
 
 /// Return the address of the contract admin
-#[allow(unused)]
 pub(crate) fn admin(store: &dyn Storage) -> Result<Addr> {
     Ok(CONFIG.load(store).map(|config| config.admin)?)
 }
