+++ contracts/service-provider-directory/src/state/mod.rs
@@ -1,0 +1,7 @@
+pub mod config;
+pub mod service_id_counter;
+pub mod services;
+
+pub(crate) use config::{admin, deposit_required, load_config, save_config, Config};
+pub(crate) use service_id_counter::next_service_id_counter;
+pub(crate) use services::services;
