+++ contracts/service-provider-directory/src/state/services.rs
@@ -1,8 +1,5 @@
-use cosmwasm_std::{Addr, Coin, Storage};
-use cw_storage_plus::{Index, IndexList, IndexedMap, Item, MultiIndex};
-use serde::{Deserialize, Serialize};
-
-use crate::error::Result;
+use cosmwasm_std::Addr;
+use cw_storage_plus::{Index, IndexList, IndexedMap, MultiIndex};
 
 use crate::types::{Service, ServiceId};
 
@@ -36,4 +33,46 @@ pub(crate) fn services<'a>() -> IndexedMap<'a, ServiceId, Service, ServiceIndex<
         ),
     };
     IndexedMap::new(SERVICES_PK_NAMESPACE, indexes)
+}
+
+#[cfg(test)]
+mod tests {
+    use cosmwasm_std::{
+        coins,
+        testing::{mock_dependencies, mock_env, mock_info},
+        Addr, Coin, Order,
+    };
+
+    use crate::{
+        msg::InstantiateMsg,
+        test_helpers::{fixture::service_fixture, helpers::get_attribute},
+        types::ServiceId,
+    };
+
+    #[test]
+    fn save_and_load_returns_a_key() {
+        let mut deps = mock_dependencies();
+        let admin = Addr::unchecked("bar");
+        let msg = InstantiateMsg {
+            admin: admin.clone(),
+            deposit_required: Coin::new(100, "unym"),
+        };
+        let info = mock_info("creator", &[]);
+
+        // Instantiate contract
+        let res = crate::instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
+        assert_eq!(res.messages.len(), 0);
+
+        // Announce
+        let msg = service_fixture().into_announce_msg();
+        let info = mock_info("anyone", &coins(100, "unym"));
+
+        let res = crate::execute(deps.as_mut(), mock_env(), info.clone(), msg.clone()).unwrap();
+        let sp_id: ServiceId = get_attribute(res.clone(), "service_id").parse().unwrap();
+        assert_eq!(sp_id, 1);
+
+        let s = super::services();
+        let k = s.keys(&deps.storage, None, None, Order::Ascending);
+        assert_eq!(k.count(), 1);
+    }
 }
