+++ contracts/service-provider-directory/src/types.rs
@@ -1,25 +1,21 @@
-use cosmwasm_std::{Addr, Coin, Storage};
-use cw_storage_plus::{Index, IndexList, IndexedMap, Item, MultiIndex};
+use cosmwasm_std::{Addr, Coin};
 use serde::{Deserialize, Serialize};
-
-use crate::error::Result;
 
 /// The directory of services are indexed by [`ServiceId`].
 pub type ServiceId = u32;
 
-/// The type of services provider supported
-#[derive(Serialize, Deserialize, Copy, Clone, PartialEq, Eq, Debug)]
-pub enum ServiceType {
-    NetworkRequester,
-}
-
-impl std::fmt::Display for ServiceType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let service_type = match self {
-            ServiceType::NetworkRequester => "network_requester",
-        };
-        write!(f, "{service_type}")
-    }
+#[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
+pub struct Service {
+    /// The address of the service.
+    pub nym_address: NymAddress,
+    /// The service type.
+    pub service_type: ServiceType,
+    /// Service owner.
+    pub owner: Addr,
+    /// Block height at which the service was added.
+    pub block_height: u64,
+    /// The deposit used to announce the service.
+    pub deposit: Coin,
 }
 
 /// The types of addresses supported.
@@ -51,17 +47,18 @@ impl ToString for NymAddress {
     }
 }
 
-#[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
-pub struct Service {
-    /// The address of the service.
-    pub nym_address: NymAddress,
-    /// The service type.
-    pub service_type: ServiceType,
-    /// Service owner.
-    pub owner: Addr,
-    /// Block height at which the service was added.
-    pub block_height: u64,
-    /// The deposit used to announce the service.
-    pub deposit: Coin,
+/// The type of services provider supported
+#[derive(Serialize, Deserialize, Copy, Clone, PartialEq, Eq, Debug)]
+pub enum ServiceType {
+    NetworkRequester,
+}
+
+impl std::fmt::Display for ServiceType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let service_type = match self {
+            ServiceType::NetworkRequester => "network_requester",
+        };
+        write!(f, "{service_type}")
+    }
 }
